Move library facets under library menu
